requirements.txt
discord
spotify
asyncio
requests
datetime
django
gunicorn
django-heroku
psutil
py-cpuinfo
